+++ newrelic/core/transaction_node.py
@@ -333,7 +333,7 @@ class TransactionNode(_TransactionNode):
                     params['userAttributes'][attr.name] = attr.value
 
             # add source context attrs for error
-            if self.settings.source_code_context.enabled:
+            if self.settings and self.settings.code_level_metrics and self.settings.code_level_metrics.enabled:
                 error.source.add_attrs(params['agentAttributes'].__setitem__)
 
             # add error specific custom params to this error's userAttributes
